[fix] Read calls had wrong return type int instead of size_t

+++ python/indexed_bzip2/indexed_bzip2.pyx
@@ -39,7 +39,7 @@ cdef extern from "indexed_bzip2/BZ2Reader.hpp":
         optional[size_t] size() except +
 
         size_t tellCompressed() except +
-        int read(int, char*, size_t) except +
+        size_t read(int, char*, size_t) except +
         bool blockOffsetsComplete() except +
         map[size_t, size_t] blockOffsets() except +
         map[size_t, size_t] availableBlockOffsets() except +
@@ -61,7 +61,7 @@ cdef extern from "indexed_bzip2/ParallelBZ2Reader.hpp":
         optional[size_t] size() except +
 
         size_t tellCompressed() except +
-        int read(int, char*, size_t) except +
+        size_t read(int, char*, size_t) except +
         bool blockOffsetsComplete() except +
         map[size_t, size_t] blockOffsets() except +
         map[size_t, size_t] availableBlockOffsets() except +
